+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 `hakyll-shortcut-links` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
+## Unreleased
+
+* [#16](https://github.com/kowainik/hakyll-shortcut-links/issues/16):
+  Support GHC-8.8.2. Update `hakyll` and `pandoc` bounds.
+  Remove support of GHC-8.6.5.
+* [#23](https://github.com/kowainik/hakyll-shortcut-links/issues/23):
+  Move CI to GitHub Actions.
+
 ## 0.0.0.0 — May 2, 2019
 
 * Initially created.